add status work

+++ customer_save.php
@@ -29,13 +29,24 @@ if($_POST[txtmode]=="add"){
     echo "<script>alert('แก้ไขข้อมูลเรียบร้อยแล้ว');</script>";
     echo "<META http-equiv='refresh' Content='0; URL=customer_show.php'> ";	
 }elseif($_GET[mode]=="del"){
-    echo "<script>alert('ลบข้อมูลเรียบร้อยแล้ว');</script>";
-    $sql="delete from tb_customer where cus_id='".$_GET[id]."'";
-    mysql_query($sql);
-    echo "<META http-equiv='refresh' Content='0; URL=customer_show.php'> ";	
-
-            
-        
     
+    $sql1="SELECT * FROM tb_general WHERE cus_id='".$_GET[id]."'";
+    $res1 = mysql_query($sql1);
+    $count1 = mysql_num_rows($res1);
+    
+    $sql2="SELECT * FROM tb_electricity WHERE cus_id='".$_GET[id]."'";
+    $res2 = mysql_query($sql2);
+    $count2 = mysql_num_rows($res2);
+
+    if($count1 or $count2 > 0){
+        echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลมีความสัมพันธ์กับตารางข้อมูลอื่น');</script>";
+        echo "<META http-equiv='refresh' Content='0; URL=customer_show.php'> ";
+    }else{
+        echo "<script>alert('ลบข้อมูลเรียบร้อยแล้ว');</script>";
+        $sql="delete from tb_customer where cus_id='".$_GET[id]."'";
+        mysql_query($sql);
+        echo "<META http-equiv='refresh' Content='0; URL=customer_show.php'> ";	
+    }
+
 }
 ?>
