+++ meter_save.php
@@ -15,10 +15,21 @@ if($_POST[txtmode]=="add"){
     echo "<script>alert('แก้ไขข้อมูลเรียบร้อยแล้ว');</script>";
     echo "<META http-equiv='refresh' Content='0; URL=meter_show.php'> ";	
 }elseif($_GET[mode]=="del"){
-    echo "<script>alert('ลบข้อมูลเรียบร้อยแล้ว');</script>";
-    $sql="delete from tb_meter where me_id='".$_GET[id]."'";
-    mysql_query($sql);
-    echo "<META http-equiv='refresh' Content='0; URL=meter_show.php'> ";	
+    
+    $sql2="SELECT * FROM tb_equipment WHERE me_id='".$_GET[id]."'";
+    $res2 = mysql_query($sql2);
+    $count2 = mysql_num_rows($res2);
+
+    if($count2 > 0){
+        echo "<script>alert('ไม่สามารถลบข้อมูลได้ เนื่องจากข้อมูลมีความสัมพันธ์กับตารางข้อมูลอื่น');</script>";
+        echo "<META http-equiv='refresh' Content='0; URL=meter_show.php'> ";
+    }else{
+        echo "<script>alert('ลบข้อมูลเรียบร้อยแล้ว');</script>";
+        $sql="delete from tb_meter where me_id='".$_GET[id]."'";
+        mysql_query($sql);
+        echo "<META http-equiv='refresh' Content='0; URL=meter_show.php'> ";
+    }
+    	
 
 }
 //elseif($_POST[txtmode]=="insert"){
